backend/requirements.txt
Django==4.2.23
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.5 
djangorestframework-simplejwt 
django-fsm==2.8.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
clamd==1.0.2
requests==2.31.0 

redis==5.0.1
django-celery-beat==2.5.0


# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
freezegun==1.2.2
pytest-asyncio==0.21.1
pytest-mock==3.12.0 

# Reporting / Export dependencies
jinja2==3.1.4
markdown==3.6
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
numpy==1.26.4
beautifulsoup4==4.12.3
python-pptx==0.6.23
drf-nested-routers==0.93.5
WeasyPrint==62.3

# S3 Storage dependencies (optional)
django-storages==1.14.2
boto3==1.34.162
